pyproject.toml
[project]
name = "dataset-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "audioread>=3.0.1",
    "beartype>=0.22.2",
    "datasets==3.6.0",
    "demucs>=4.0.1",
    "gradio-utils",
    "librosa>=0.11.0",
    "llama-cpp-python==0.2.64",
    "matchering>=2.0.6",
    "onnx>=1.19.1",
    "onnx2pytorch>=0.5.3",
    "pyannote-audio>=3.1.1,<4.0", # v4.0+ requires torchcodec which has std::bad_alloc bug on WSL2
    "pydub>=0.25.1",
    "pyrubberband>=0.4.0",
    "pysrt>=1.1.2",
    "pytorch-lightning>=2.5.0",
    "rapidfuzz>=3.13.0",
    "requests==2.32.3",
    "rotary-embedding-torch>=0.8.9",
    "soundfile>=0.13.1",
    "starlette==0.38.2",
    "torch[cu128]==2.8.0",
    "urllib3==2.3",
    "psutil>=6.1.0",
    "optimum>=2.0.0",
    "whisperx>=3.7.0",
    "onnxruntime-gpu==1.20.1",
    "nvidia-cudnn-cu12>=9.1.0",
]

[tool.uv]
index-strategy = "unsafe-best-match"
override-dependencies = [
    "numpy<2.0",
    "pydantic<=2.10.6",
    "gradio==4.44.1",
    "urllib3>=2.0",
    "pyannote-audio>=3.1.1,<4.0",  # v4.0+ requires torchcodec
    "onnxruntime-gpu==1.20.1",
    "pytorch-lightning>=2.5.0",
    "ctranslate2>=4.6.0",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "triton>=3.3.0"]

[tool.uv.sources]
gradio-utils = { git = "https://github.com/JarodMica/gradio_utils.git" }
whisperx = { git = "https://github.com/m-bain/whisperX.git", branch = "main" }
torch = [
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cu128", extra = "cu128" },
]
torchaudio = [
  { index = "pytorch-cu128", extra = "cu128" },
]
triton = [
  { index = "pytorch-cu128" },
]


[project.optional-dependencies]
cu128 = [
  "torch==2.8.0",
  "torchvision==0.23.0",
  "torchaudio==2.8.0"
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
